set Results as generic data attribute in kaleido parser
